fix create_test_block
